+++ src/main/java/com/eugeniokg/corebankingledger/transaction/EntryType.java
@@ -1,0 +1,6 @@
+package com.eugeniokg.corebankingledger.transaction;
+
+public enum EntryType {
+    DEBIT,
+    CREDIT
+}